Финальная версия. gl&hf

+++ LINK_obrnadzor.py
@@ -17,16 +17,21 @@ def GetSoup():
     return soup
 
 
-def InitDbConnectionObrnadzor(option):
+def InitDbConnectionObrnadzor(option, option_delete):
     '''
-    option == True => подключение к локальной бд
+    option == True => подключение к локальной бд \n
     option == False => подключение к бд на VPS
+
+    option_delete == True => удаление коллекции \n
+    option_delete == False => коллекция не удаляется
 
     Подключаемся к коллекции obrnadzor, из которой будет происходить выборка по критериям
     '''
     if option: # Подключение к локальной, пересоздание и обновление коллекции obrnadzor_local
         client = MongoClient('localhost', 27017)
-        db = client.RKN
+        db = client.RKNN
+        if 'obrnadzor_local' in db.collection_names() and option_delete:
+            db.drop_collection('obrnadzor_local')
         collection = db.obrnadzor_local
         return collection
     else: # Подключение к бд на сервере, пересоздание и обновление коллекции obrnadzor
@@ -35,22 +40,27 @@ def InitDbConnectionObrnadzor(option):
                     password='MW6Vh6dlw4FaNv0aSi4Rs15Y',
                     authSource='admin',
                     authMechanism='SCRAM-SHA-1')
-        db = client.RKN
+        db = client.RKNN
+        if 'obrnadzor' in db.collection_names() and option_delete:
+            db.drop_collection('obrnadzor')
         collection = db.obrnadzor
         return collection
 
     
-def InitDbConnectionEdu(option):
+def InitDbConnectionEdu(option, option_delete):
     '''
-    option == True => подключение к локальной бд
+    option == True => подключение к локальной бд \n
     option == False => подключение к бд на VPS
+
+    option_delete == True => удаление коллекции \n
+    option_delete == False => коллекция не удаляется
 
     Подключаемся к коллекции edu, в которую будет производиться запись данных из obrnadzor
     '''
     if option: # Подключение к локальной, пересоздание и обновление коллекции edu_records
         client = MongoClient('localhost', 27017)
-        db = client.RKN
-        if 'edu_records' in db.collection_names():
+        db = client.RKNN
+        if 'edu_records' in db.collection_names() and option_delete:
             db.drop_collection('edu_records')
         collection = db.edu_records
         return collection
@@ -60,8 +70,8 @@ def InitDbConnectionEdu(option):
                     password='MW6Vh6dlw4FaNv0aSi4Rs15Y',
                     authSource='admin',
                     authMechanism='SCRAM-SHA-1')
-        db = client.RKN
-        if 'edu' in db.collection_names():
+        db = client.RKNN
+        if 'edu' in db.collection_names() and option_delete:
             db.drop_collection('edu')
         collection = db.edu
         return collection
@@ -85,23 +95,29 @@ def GetRegionsDict():
     return dictionary
 
 
-def InsertIntoDb(option_obrnadzor, option_edu, inn = None, region = None):
+def InsertIntoDb(option_obrnadzor, option_edu, option_delete_obrnadzor, option_delete_edu, inn = None, region = None):
     '''
-    option_obrnadzor == True => подключение к локальной бд (obrnadzor_local)
+    option_obrnadzor == True => подключение к локальной бд (obrnadzor_local) \n
     option_obrnadzor == False => подключение к бд на VPS (obrnadzor)
 
-    option_edu == True => подключение к локальной бд (edu_local)
-    option_edu == False => подключение к бд на VPS (edu)
+    option_edu == True => подключение к локальной бд (edu_local) \n
+    option_edu == False => подключение к бд на VPS (edu) \n
 
-    Выборка элементов из коллекции obrnadzor и вставка их в edu
+    option_delete_obrnadzor == True => удаление коллекции \n
+    option_delete_obrnadzor == False => коллекция не удаляется
+
+    option_delete_edu == True => удаление коллекции \n
+    option_delete_edu == False => коллекция не удаляется
+
+    Выборка элементов из коллекции obrnadzor и вставка их в edu.
     Выборка происходит по 2-м критериям:
         1) ИНН
         2) Номер региона
-    Можно ввести как один критерий, так и сразу два
-    Если не вводить никаких критериев отбора, то будут выбраны и вставлены все элементы из коллекции
+    Можно ввести как один критерий, так и сразу два. \n
+    Если не вводить никаких критериев отбора, то будут выбраны и вставлены все элементы из коллекции.
     '''
-    collection = InitDbConnectionObrnadzor(option_obrnadzor)
-    collection_edu = InitDbConnectionEdu(option_edu)
+    collection = InitDbConnectionObrnadzor(option_obrnadzor, option_delete_obrnadzor)
+    collection_edu = InitDbConnectionEdu(option_edu, option_delete_edu) 
     dictionary = GetRegionsDict()
     query = {}
     if region != None and inn == None:
@@ -116,4 +132,4 @@ def InsertIntoDb(option_obrnadzor, option_edu, inn = None, region = None):
 
 
 if __name__ == '__main__':
-    InsertIntoDb(True, False, None, None)
+    InsertIntoDb(True, False, False, True, None, None)
